Simplify route marshaling logic
`Zend\Expressive\Route` _always_ creates a non-empty route name, and
this same name will be used to populate the underlying Aura route
collector. As such, we can match only on that, and always be successful.

To simulate this best, I updated `testSuccessfulRouteResultComposesMatchedRoute()`
to _not_ mock any dependencies, to ensure that the Aura route created is
based off of the `Route` passed to the router; prior to the changes,
they were out-of-sync with regards to the name.