Fixed two bugs discovered when integrating in zend-expressive
The first issue observed was that when multiple routes are injected with
the same path, but different allowed HTTP methods, the value reported in
the `Allow` header only reported the HTTP method(s) supported by the
first matching route, and not the full list of supported methods. The
solution to this is to prefer the values in the `$pathMethodMap` over
using the failed route's `$allows` property. (I've also updated the code
to pass the value of merging `$allowedMethods` through `array_unique()`
to ensure no duplicates occur.)

The second issue observed was that if a root path matches the URI, but
no subpath does, we were returning a method failure (405), and not a
not found (404). This was resolved by looking in the failed route's
`$failedRule`; if it is equivalent to `Aura\Router\Rule\Path`, then we
have a 404 condition.

These changes were due to the upgrade to Aura.Router v3, and are
unrelated to the changes to support HEAD or OPTIONS.