Ensure that when no HTTP methods are present non-implicit methods result in a 405
Discovered when doing integration tests from #398. Essentially, Aura
successfully matches a path if no methods are specified in the original
route, when it should treat that as a 405 failure. We now check for an
empty $route->allows array on a successful match, and, if so, treat it
as a potential failure (potential, as HEAD and OPTIONS will be
implicitly supported).