Ensure router works as expected under zend-expressive-router v3
`RouteResult::fromRouteFailure()` results in a few internal changes that
can raise issues with Aura.Router due to the fact that Aura.Router
treats an empty set of allowed methods as EITHER allowing all HTTP
methods or none.

Additionally, `RouteResult::getAllowedMethods()` now returns `['*']`
when all methods are allowed, which is a valid string for inclusion in
an `Allows` header to indicate any method is allowed. This change
required some minor updates to test expectations.